+++ lib/fog/azurerm/requests/traffic_manager/create_traffic_manager_endpoint.rb
@@ -1,25 +1,25 @@
 module Fog
   module TrafficManager
     class AzureRM
-      # Real class for TrafficManager Request
+      # Real class for Traffic Manager Request
       class Real
-        def create_traffic_manager_endpoint(resource_group, name, traffic_manager_profile_name, type, target_resource_id,
-                                            target, weight, priority, endpoint_location, min_child_endpoints)
-          msg = "Creating Traffic Manager Endpoint: #{name}."
+        def create_traffic_manager_endpoint(endpoint_hash)
+          msg = "Creating Traffic Manager Endpoint: #{endpoint_hash[:name]}."
           Fog::Logger.debug msg
-          endpoint_parameters = get_endpoint_parameters(target_resource_id, target, weight, priority, endpoint_location, min_child_endpoints)
+          endpoint_parameters = get_endpoint_object(endpoint_hash[:target_resource_id], endpoint_hash[:target], endpoint_hash[:weight], endpoint_hash[:priority], endpoint_hash[:endpoint_location], endpoint_hash[:min_child_endpoints])
           begin
-            traffic_manager_endpoint = @traffic_mgmt_client.endpoints.create_or_update(resource_group, traffic_manager_profile_name,
-                                                                                       type, name, endpoint_parameters)
+            traffic_manager_endpoint = @traffic_mgmt_client.endpoints.create_or_update(endpoint_hash[:resource_group], endpoint_hash[:traffic_manager_profile_name],
+                                                                                       endpoint_hash[:type], endpoint_hash[:name], endpoint_parameters)
           rescue MsRestAzure::AzureOperationError => e
             raise_azure_exception(e, msg)
           end
+          Fog::Logger.debug "Traffic Manager Endpoint: #{endpoint_hash[:name]} created successfully."
           traffic_manager_endpoint
         end
 
         private
 
-        def get_endpoint_parameters(target_resource_id, target, weight, priority, endpoint_location, min_child_endpoints)
+        def get_endpoint_object(target_resource_id, target, weight, priority, endpoint_location, min_child_endpoints)
           endpoint = Azure::ARM::TrafficManager::Models::Endpoint.new
           endpoint.min_child_endpoints = min_child_endpoints
           endpoint.target_resource_id = target_resource_id
@@ -33,20 +33,20 @@ module Fog
 
       # Mock class for TrafficManager Request
       class Mock
-        def create_traffic_manager_endpoint(resource_group, name, traffic_manager_profile_name, type, target_resource_id,
-                                            target, weight, priority, endpoint_location, min_child_endpoints)
-          response = {}
-          properties = {}
-          properties['weight'] = weight
-          properties['priority'] = priority
-          properties['targetResourceId'] = target_resource_id unless target_resource_id.nil?
-          properties['target'] = target unless target.nil?
-          properties['endpointLocation'] = endpoint_location unless endpoint_location.nil?
-          properties['minChildEndpoints'] = min_child_endpoints unless min_child_endpoints.nil?
-          response['id'] = "/subscriptions/######/resourceGroups/#{resource_group}/providers/Microsoft.Network/trafficManagerProfiles/#{traffic_manager_profile_name}/#{type}Endpoints/#{name}?api-version=2015-11-01"
-          response['type'] = "Microsoft.Network/trafficManagerProfiles/#{type}Endpoints"
-          response['properties'] = properties
-          response
+        def create_traffic_manager_endpoint(*)
+          endpoint = {
+            'name' => '{endpoint-name}',
+            'type' => 'Microsoft.Network/trafficManagerProfiles/externalEndpoints',
+            'properties' => {
+              'target' => 'myendpoint.contoso.com',
+              'endpointStatus' => 'Enabled',
+              'weight' => 10,
+              'priority' => 5,
+              'endpointLocation' => 'northeurope'
+            }
+          }
+          endpoint_mapper = Azure::ARM::TrafficManager::Models::Endpoint.mapper
+          @traffic_mgmt_client.deserialize(endpoint_mapper, endpoint, 'result.body')
         end
       end
     end